+++ src/Managers/StateKeys.ts
@@ -1,0 +1,8 @@
+/*VsCode Stuff*/
+export enum StateKeys
+{
+    instance,
+    url,
+    user,
+    password
+}
