+++ src/ServiceNow/IsysRecord.ts
@@ -1,0 +1,13 @@
+import {Relation} from './Relation';
+
+//Interface declaring fields used from sys_metadata
+export interface IsysRecord
+{
+    sys_class_name: string;
+    sys_id: string;
+    sys_policy: string;
+    sys_updated_on: Date;
+    sys_created_on: Date;
+    sys_package: Relation;
+    sys_scope: Relation;
+}
