+++ src/ServiceNow/IsysRelation.ts
@@ -1,0 +1,5 @@
+export interface IsysRelation
+{
+    link: string;
+    display_value: string;
+}
